services/api-gateway/requirements.txt
# Athena API Gateway 依赖包
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
aiofiles==23.2.1
pydantic==2.5.0
PyYAML==6.0.1
python-multipart==0.0.6
PyJWT==2.8.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0